[fix] (litecoin) move over testnet tests to mainnet, remove testnet server
