+++ packages/litecoin-payments/test/e2e.mainnet.test.ts
@@ -5,11 +5,11 @@ import { FeeRateType, BalanceResult, TransactionStatus, NetworkType } from '@faa
 
 import {
   HdLitecoinPayments, LitecoinTransactionInfo, HdLitecoinPaymentsConfig,
-  LitecoinSignedTransaction, LitecoinUnsignedTransaction, AddressType,
+  LitecoinSignedTransaction, SinglesigAddressType, AddressType,
 } from '../src'
 
-import { txInfo_e10d7, signedTx_valid, signedTx_invalid } from './fixtures/transactions'
-import { accountsByAddressType, AccountFixture } from './fixtures/accounts'
+import { txInfo_e10d7 } from './fixtures'
+import fixtures from './fixtures/singlesigMainnet'
 import { END_TRANSACTION_STATES, delay, expectEqualWhenTruthy, logger, expectEqualOmit } from './utils'
 import { toBigNumber } from '@faast/ts-common'
 import BigNumber from 'bignumber.js'
@@ -32,6 +32,12 @@ if (fs.existsSync(secretXprvFilePath)) {
     `File ${SECRET_XPRV_FILE} missing. Send and sweep e2e mainnet tests will be skipped. To enable them ask Dylan to share the file with you.`,
   )
 }
+
+const addressTypesToTest: SinglesigAddressType[] = [
+  AddressType.Legacy,
+  AddressType.SegwitP2SH,
+  AddressType.SegwitNative,
+]
 
 function assertTxInfo(actual: LitecoinTransactionInfo, expected: LitecoinTransactionInfo): void {
   expectEqualOmit(actual, expected, ['data.confirmations', 'confirmations'])
@@ -240,110 +246,140 @@ describeAll('e2e mainnet', () => {
     expect(signedTx.data.unsignedTxHash).toMatch(/^[a-f0-9]+$/)
   })
 
-  async function pollUntilEnded(signedTx: LitecoinSignedTransaction) {
-    const txId = signedTx.id
-    logger.log('polling until ended', txId)
-    let tx: LitecoinTransactionInfo | undefined
-    while (!testsComplete && (!tx || !END_TRANSACTION_STATES.includes(tx.status) || tx.confirmations === 0)) {
-      try {
-        tx = await payments.getTransactionInfo(txId)
-      } catch (e) {
-        if (e.message.includes('Transaction not found')) {
-          logger.log('tx not found yet', txId, e.message)
-        } else {
-          throw e
-        }
-      }
-      await delay(5000)
-    }
-    if (!tx) {
-      throw new Error(`failed to poll until ended ${txId}`)
-    }
-    logger.log(tx.status, tx)
-    expect(tx.id).toBe(signedTx.id)
-    expect(tx.fromAddress).toBe(signedTx.fromAddress)
-    expectEqualWhenTruthy(tx.fromExtraId, signedTx.fromExtraId)
-    expect(tx.toAddress).toBe(signedTx.toAddress)
-    expectEqualWhenTruthy(tx.toExtraId, signedTx.toExtraId)
-    expect(tx.data).toBeDefined()
-    expect(tx.status).toBe(TransactionStatus.Confirmed)
-    expect(tx.isConfirmed).toBe(true)
-    expect(tx.isExecuted).toBe(true)
-    expect(tx.confirmationId).toMatch(/^\w+$/)
-    expect(tx.confirmationTimestamp).toBeDefined()
-    expect(tx.confirmations).toBeGreaterThan(0)
-    return tx
-  }
+  for (let addressType of addressTypesToTest) {
+    const { xpub, addresses, sweepTxSize } = fixtures[addressType]
 
-  jest.setTimeout(300 * 1000)
-
-  it.skip('end to end sweep', async () => {
-    const indicesToTry = [0, 1]
-    const balances: { [i: number]: BalanceResult } = {}
-    let indexToSweep: number = -1
-    for (const index of indicesToTry) {
-      const balanceResult = await payments.getBalance(index)
-      balances[index] = balanceResult
-      if (balanceResult.sweepable) {
-        indexToSweep = index
-        break
+    describe(addressType, () => {
+      const paymentsConfig: HdLitecoinPaymentsConfig = {
+        hdKey: secretXprv,
+        network: NetworkType.Mainnet,
+        addressType,
+        logger,
+        minChange: '0.01',
+        targetUtxoPoolSize: 5,
       }
-    }
-    if (indexToSweep < 0) {
-      const allAddresses = await Promise.all(indicesToTry.map(async i => (await payments.getPayport(i)).address))
-      throw new Error(`Cannot end to end test sweeping due to lack of funds. Send mainnet LTC to any of the following addresses and try again. ${JSON.stringify(allAddresses)}`)
-    }
-    const recipientIndex = indexToSweep === indicesToTry[0] ? indicesToTry[1] : indicesToTry[0]
-    try {
-      const unsignedTx = await payments.createSweepTransaction(indexToSweep, recipientIndex, { feeRate, feeRateType })
-      const signedTx = await payments.signTransaction(unsignedTx)
-      logger.log(`Sweeping ${signedTx.amount} from ${indexToSweep} to ${recipientIndex} in tx ${signedTx.id}`)
-      expect(await payments.broadcastTransaction(signedTx)).toEqual({
-        id: signedTx.id,
+      const payments = new HdLitecoinPayments(paymentsConfig)
+      it('get correct xpub', async () => {
+        expect(payments.xpub).toEqual(xpub)
       })
-      const tx = await pollUntilEnded(signedTx)
-      expect(tx.amount).toEqual(signedTx.amount)
-      expect(tx.fee).toEqual(signedTx.fee)
-    } catch (e) {
-      if ((e.message || (e as string)).includes('balance is not sufficient')) {
-        logger.log('Ran consecutive tests too soon, previous sweep not complete. Wait a minute and retry')
+      for (let iStr in addresses) {
+        const i = Number.parseInt(iStr)
+        it(`get correct address for index ${i}`, async () => {
+          expect(await payments.getPayport(i)).toEqual({ address: (addresses as any)[i] })
+        })
       }
-      throw e
-    }
-  })
-  
-  it('end to end send', async () => {
-    const indicesToTry = [1, 0]
-    const balances: { [i: number]: BalanceResult } = {}
-    let indexToSend: number = -1
-    let highestBalance = toBigNumber(0)
-    for (const index of indicesToTry) {
-      const balanceResult = await payments.getBalance(index)
-      balances[index] = balanceResult
-      if (toBigNumber(balanceResult.confirmedBalance).gt(highestBalance)) {
-        indexToSend = index
-        break
+
+      async function pollUntilEnded(signedTx: LitecoinSignedTransaction) {
+        const txId = signedTx.id
+        logger.log('polling until ended', txId)
+        let tx: LitecoinTransactionInfo | undefined
+        while (!testsComplete && (!tx || !END_TRANSACTION_STATES.includes(tx.status) || tx.confirmations === 0)) {
+          try {
+            tx = await payments.getTransactionInfo(txId)
+          } catch (e) {
+            if (e.message.includes('Transaction not found')) {
+              logger.log('tx not found yet', txId, e.message)
+            } else {
+              throw e
+            }
+          }
+          await delay(5000)
+        }
+        if (!tx) {
+          throw new Error(`failed to poll until ended ${txId}`)
+        }
+        logger.log(tx.status, tx)
+        expect(tx.id).toBe(signedTx.id)
+        expect(tx.fromAddress).toBe(signedTx.fromAddress)
+        expectEqualWhenTruthy(tx.fromExtraId, signedTx.fromExtraId)
+        expect(tx.toAddress).toBe(signedTx.toAddress)
+        expectEqualWhenTruthy(tx.toExtraId, signedTx.toExtraId)
+        expect(tx.data).toBeDefined()
+        expect(tx.status).toBe(TransactionStatus.Confirmed)
+        expect(tx.isConfirmed).toBe(true)
+        expect(tx.isExecuted).toBe(true)
+        expect(tx.confirmationId).toMatch(/^\w+$/)
+        expect(tx.confirmationTimestamp).toBeDefined()
+        expect(tx.confirmations).toBeGreaterThan(0)
+        return tx
       }
-    }
-    if (indexToSend < 0) {
-      const allAddresses = await Promise.all(indicesToTry.map(async i => (await payments.getPayport(i)).address))
-      throw new Error(`Cannot end to end test sweeping due to lack of funds. Send LTC to any of the following addresses and try again. ${JSON.stringify(allAddresses)}`)
-    }
-    const recipientIndex = indexToSend === indicesToTry[0] ? indicesToTry[1] : indicesToTry[0]
-    const unsignedTx = await payments.createTransaction(
-      indexToSend,
-      recipientIndex,
-      '0.05',
-      { useUnconfirmedUtxos: true }, // Prevents consecutive tests from failing
-    )
-    const signedTx = await payments.signTransaction(unsignedTx)
-    logger.log(`Sending ${signedTx.amount} from ${indexToSend} to ${recipientIndex} in tx ${signedTx.id}`)
-    expect(await payments.broadcastTransaction(signedTx)).toEqual({
-      id: signedTx.id,
+
+      jest.setTimeout(300 * 1000)
+
+      it.skip('end to end sweep', async () => {
+        const indicesToTry = [5, 6]
+        const balances: { [i: number]: BalanceResult } = {}
+        let indexToSweep: number = -1
+        for (const index of indicesToTry) {
+          const balanceResult = await payments.getBalance(index)
+          balances[index] = balanceResult
+          if (balanceResult.sweepable) {
+            indexToSweep = index
+            break
+          }
+        }
+        if (indexToSweep < 0) {
+          const allAddresses = await Promise.all(indicesToTry.map(async i => (await payments.getPayport(i)).address))
+          throw new Error(`Cannot end to end test sweeping due to lack of funds. Send LTC to any of the following addresses and try again. ${JSON.stringify(allAddresses)}`)
+        }
+        const recipientIndex = indexToSweep === indicesToTry[0] ? indicesToTry[1] : indicesToTry[0]
+        const satPerByte = 44
+        const unsignedTx = await payments.createSweepTransaction(indexToSweep, recipientIndex, {
+          feeRate: satPerByte.toString(),
+          feeRateType: FeeRateType.BasePerWeight,
+          useUnconfirmedUtxos: true, // Prevents consecutive tests from failing
+        })
+        const signedTx = await payments.signTransaction(unsignedTx)
+        expect(signedTx.inputUtxos).toBeDefined()
+        expect(signedTx.inputUtxos!.length).toBe(1)
+        expect(signedTx.externalOutputs).toBeDefined()
+        expect(signedTx.externalOutputs!.length).toBe(1)
+        const feeNumber = new BigNumber(signedTx.fee).toNumber()
+        expect(feeNumber).toBe((sweepTxSize*satPerByte)*1e-8)
+        logger.log(`Sweeping ${signedTx.amount} from ${indexToSweep} to ${recipientIndex} in tx ${signedTx.id}`)
+        expect(await payments.broadcastTransaction(signedTx)).toEqual({
+          id: signedTx.id,
+        })
+        const tx = await pollUntilEnded(signedTx)
+        // const tx = await payments.getTransactionInfo(signedTx.id)
+        expect(tx.amount).toEqual(signedTx.amount)
+        expect(tx.fee).toEqual(signedTx.fee)
+      })
+
+      it.skip('end to end send', async () => {
+        const indicesToTry = [7, 8]
+        const balances: { [i: number]: BalanceResult } = {}
+        let indexToSend: number = -1
+        let highestBalance = toBigNumber(0)
+        for (const index of indicesToTry) {
+          const balanceResult = await payments.getBalance(index)
+          balances[index] = balanceResult
+          if (toBigNumber(balanceResult.confirmedBalance).gt(highestBalance)) {
+            indexToSend = index
+            break
+          }
+        }
+        if (indexToSend < 0) {
+          const allAddresses = await Promise.all(indicesToTry.map(async i => (await payments.getPayport(i)).address))
+          throw new Error(`Cannot end to end test sweeping due to lack of funds. Send LTC to any of the following addresses and try again. ${JSON.stringify(allAddresses)}`)
+        }
+        const recipientIndex = indexToSend === indicesToTry[0] ? indicesToTry[1] : indicesToTry[0]
+        const unsignedTx = await payments.createTransaction(
+          indexToSend,
+          recipientIndex,
+          '0.0001',
+          { useUnconfirmedUtxos: true }, // Prevents consecutive tests from failing
+        )
+        const signedTx = await payments.signTransaction(unsignedTx)
+        logger.log(`Sending ${signedTx.amount} from ${indexToSend} to ${recipientIndex} in tx ${signedTx.id}`)
+        expect(await payments.broadcastTransaction(signedTx)).toEqual({
+          id: signedTx.id,
+        })
+        const tx = await pollUntilEnded(signedTx)
+        // const tx = await payments.getTransactionInfo(signedTx.id)
+        expect(tx.amount).toEqual(signedTx.amount)
+        expect(tx.fee).toEqual(signedTx.fee)
+      })
     })
-    const tx = await pollUntilEnded(signedTx)
-    //const tx = await payments.getTransactionInfo(signedTx.id)
-    expect(tx.amount).toEqual(signedTx.amount)
-    expect(tx.fee).toEqual(signedTx.fee)
-  })
+  }
 })
