+++ packages/litecoin-payments/test/fixtures/singlesigMainnet.ts
@@ -1,0 +1,36 @@
+import { AddressType } from '../../src'
+export default {
+  [AddressType.Legacy]: {
+    xpub: 'xpub6BvtsVTfEJWNuwLGduLMBYK9oBBpYmteSoayxvXMMjNMqWsbhX4xNJeTL83BLoLNZoqAdJhr86JHqEJA16ZSBfQG9cEkx3d1MyHrZSSV8X3',
+    sweepTxSize: 192,
+    addresses: {
+      0: 'Lb2JxhU6gEXZWuKSzDa8KmrbJuCq5Zp3WM',
+      5: 'LYo7yxSz3fYRntwSigRXfnvxSgJkMXfnrk',
+      6: 'LeBPwcea5y4PSD9eb4mgXQaxG9XfZgqgCb',
+      7: 'LgAp5DNiqb9FdTKbGZT7UF2m959iYbSQWz',
+      8: 'LbLsqrRe9o8edvHiPxgmbDF57CbUuXcPre',
+    },
+  },
+  [AddressType.SegwitP2SH]: {
+    xpub: 'xpub6CrM2pH5CVqaBb4eDGBd1c3bVVgHb9PXzJXqKqjrmyaPR8Jbv7QeCsSzH1pRVNhfdx24V9AKYD1g9ze6KhMjXoLY1qT8g1DSL4bEyEnUsiR',
+    sweepTxSize: 133,
+    addresses: {
+      0: 'MS4UPbL5FBbWtzLb4QB6SyeS96MbnpJpus',
+      5: 'MHrZeWbuYk9FyxNSv1xWPi1HJghQwsQZ3Z',
+      6: 'MWqbvRkZH1nV31ToqMHxKaB51rs97qZetK',
+      7: 'MViazG1Uf29ymt2X3fLrvZCziVPinyj7wg',
+      8: 'MCMcRqdaXXMpYmea5PRKZH2xZia7FYHV13',
+    },
+  },
+  [AddressType.SegwitNative]: {
+    xpub: 'xpub6CrMcKhbvSyc3ciFxZ4TYkdexCsKCA3hQVCYzn6UJHUA5GHkEzUt3w72kGrQGpXdwR4LHc5JGGoqEyq6FX3MD18oujhe4AAqXh6veaLF8XZ',
+    sweepTxSize: 109,
+    addresses: {
+      0: 'ltc1q9ek9srkxa69l8p9qdk8v2ntzs9vetxnr6xhvf4',
+      5: 'ltc1q46zmg3v87rr942cad3hlv5ghusz79pdtzpazzh',
+      6: 'ltc1qrx0dulccq3w44jr8q4m9492s2jju7vl65wdmaa',
+      7: 'ltc1qn2788flds3hay5w7a2cj3z9ce2p4du4vn5aedp',
+      8: 'ltc1qn862r8z2dkez0m8nl5zaw0t7uvq448p78m7cwr',
+    },
+  }
+}
